add reference links to bioc/r implementations
